+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Updated Event handling for Runner related events
 - Updated Event handling for Batch related events
 - Updated Event handling for running the AutoTransform script
+- Add EventNotifier components to replace env variable EventHandler replacement
 
 ### New Components
 - ScriptInput that uses a script to generate a list of Items